+++ backup_codes/apps.py
@@ -1,4 +1,5 @@
 from django.apps import AppConfig
+from portal import container
 
 
 class BackupCodesConfig(AppConfig):
@@ -6,3 +7,12 @@ class BackupCodesConfig(AppConfig):
 
     def ready(self):
         import backup_codes.signals  # noqa
+
+        """
+        Wire the dependency-injector.
+        Each application must add this code in the apps.py file
+        and list each module that uses dependency-injections
+        """
+        from . import forms
+
+        container.wire(modules=[forms])